Update README.md
Corrected a `toml` formatting issue. Need to create another project board for administrative tasks related to the deleted "Installation" section. Obviously crates.io and pypi.org make excellent candidates for distributing the software, but I would like to explore the steps needed to register `edm` with an APT repository as well as with Snap.